+++ cmd/portugol/format_test.go
@@ -118,7 +118,6 @@ func TestFormatterPreservesMalformedFiles(t *testing.T) {
 		{"single_quoted_string", "../../testdata/conformance/visualg-3.0.7/probes/single-quoted-string/source.alg"},
 		{"alias-caracter-callable", "../../testdata/conformance/visualg-3.0.7/probes/alias-caracter-callable/source.alg"},
 		{"header-name-without-keyword", "../../testdata/conformance/visualg-3.0.7/probes/header-name-without-keyword/source.alg"},
-		{"brace-comment-in-expression", "../../testdata/conformance/visualg-3.0.7/probes/brace-comment-in-expression/source.alg"},
 	} {
 		t.Run(tt.name, func(t *testing.T) {
 			src, err := os.ReadFile(tt.path)